# pulse-query-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hospital Analytics Platform",
    "description": "Backend API for the Hospital Analytics Platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/templates/": {
      "get": {
        "tags": ["templates"],
        "summary": "List Templates",
        "description": "Retrieve a list of available widget templates with filtering and pagination.\n\nArgs:\n    db (AsyncSession): Database session dependency.\n    category (Optional[str]): Filter templates by category name.\n    search (Optional[str]): Case-insensitive search term for title or description.\n    limit (int): Maximum number of records to return.\n    current_user (User): Authenticated user dependency.\n\nReturns:\n    List[WidgetTemplate]: Filtered list of templates.",
        "operationId": "list_templates_api_v1_templates__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter templates by category",
              "title": "Category"
            },
            "description": "Filter templates by category"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search title or description",
              "title": "Search"
            },
            "description": "Search title or description"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max records to return",
              "default": 30,
              "title": "Limit"
            },
            "description": "Max records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateResponse"
                  },
                  "title": "Response List Templates Api V1 Templates  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["templates"],
        "summary": "Create Template",
        "description": "Register a new analytics template in the system.\n\nArgs:\n    template_in (TemplateCreate): Validated template payload.\n    db (AsyncSession): Database session.\n    current_user (User): Authenticated user.\n\nReturns:\n    WidgetTemplate: The persisted template object.",
        "operationId": "create_template_api_v1_templates__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/{template_id}": {
      "get": {
        "tags": ["templates"],
        "summary": "Get Template",
        "description": "Retrieve details for a specific template.\n\nArgs:\n    template_id (UUID): The unique ID of the template.\n    db (AsyncSession): Database session.\n    current_user (User): Authenticated user.\n\nReturns:\n    WidgetTemplate: The requested template.\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "get_template_api_v1_templates__template_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["templates"],
        "summary": "Update Template",
        "description": "Update an existing template's definition.\n\nArgs:\n    template_id (UUID): The unique ID of the template.\n    template_in (TemplateUpdate): Partial update payload.\n    db (AsyncSession): Database session.\n    current_user (User): Authenticated user.\n\nReturns:\n    WidgetTemplate: The updated template.\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "update_template_api_v1_templates__template_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["templates"],
        "summary": "Delete Template",
        "description": "Hard delete a template from the registry.\n\nArgs:\n    template_id (UUID): The unique ID of the template.\n    db (AsyncSession): Database session.\n    current_user (User): Authenticated user.\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "delete_template_api_v1_templates__template_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TemplateCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sql_template": {
            "type": "string",
            "title": "Sql Template",
            "description": "SQL with {{handlebars}} placeholders"
          },
          "category": {
            "type": "string",
            "minLength": 2,
            "title": "Category"
          },
          "parameters_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters Schema",
            "description": "JSON Schema definition for dynamic form generation."
          }
        },
        "type": "object",
        "required": ["title", "sql_template", "category"],
        "title": "TemplateCreate",
        "description": "Payload for creating a new template.\nInherits validation from TemplateBase."
      },
      "TemplateResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sql_template": {
            "type": "string",
            "title": "Sql Template",
            "description": "SQL with {{handlebars}} placeholders"
          },
          "category": {
            "type": "string",
            "minLength": 2,
            "title": "Category"
          },
          "parameters_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters Schema",
            "description": "JSON Schema definition for dynamic form generation."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": ["title", "sql_template", "category", "id"],
        "title": "TemplateResponse",
        "description": "API Response model for a Template.\nInclude database ID."
      },
      "TemplateUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sql_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql Template"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "parameters_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters Schema"
          }
        },
        "type": "object",
        "title": "TemplateUpdate",
        "description": "Payload for updating an existing template.\nPartial fields allowed."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}